Ensure that codec tests are run (#1763)
* Move test_codec.py to the right directory
* Call PyObjectConversions.Reset from Python via a proxy to keep it internal
* Sign Python.Test DLL and make Python.Runtime internals visible for it
